CIRCULAR IMPORT: change fillMasterThreadInput to virtual function in IChimeraWindow
Also fixed the circulat import problem with vs tools https://stackoverflow.com/questions/1137966/displaying-the-include-hierarchy-for-a-c-file-in-visual-studio
while doing that, also try to implement the forward declarition as much as possible so that the code may complie faster?
Also remove the PlotCtrl from the dependence, but still in the project.
Also refactored some files to avoid circular import