TO BE TESTED IN EXP: add threadsafe queue and restructure how andor pic is transfered between different worker thread
This comes with no non-deterministic delay as in the Qt or MFC way. Thq threadsafe queue is automatically doing the wait/wake feature among differetn thread.
The picture is read in the grabber thread and the original worker thraed will acquire the image buffer and requeue a new one with the circular buffer, which allows fastest image retrieval time without any significant delay for the next iamge to be read in the meantime.
Still need to be tested on the real system